# Add some timing and test count to karma output.