@tmcw 2015-01-09
:+1: definitely better than the current behavior.